requirements.txt
graphene-django==2.15.0
django==3.2.13
pyjwt==1.7.1
pillow==9.3.0
graphene-file-upload==1.2.2
mysqlclient==2.0.3
django-environ==0.7.0
wsgicors==0.7.0